requirements.txt
absl-py==1.4.0
accelerate==1.4.0
addict==2.4.0
aiohappyeyeballs==2.5.0
aiohttp==3.11.13
aiosignal==1.3.2
annotated-types==0.7.0
anthropic==0.49.0
anyio==3.7.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.2.1
async-timeout==5.0.1
attrs==25.1.0
babel==2.17.0
backcall==0.2.0
beautifulsoup4==4.13.3
binaryornot==0.4.4
bleach==6.2.0
blinker==1.9.0
boto3==1.17.112
botocore==1.20.112
cachetools==5.3.1
certifi==2023.7.22
cffi==1.17.1
chardet==5.2.0
charset-normalizer==2.0.12
click==8.1.6
cmake==3.27.0
comm==0.2.2
configparser==6.0.0
contourpy==1.1.0
cookiecutter==2.6.0
cryptography==3.4.7
cycler==0.11.0
Cython==0.29.30
datasets==2.19.1
debugpy==1.6.7
decorator==5.1.1
deepspeed==0.15.4
defusedxml==0.7.1
dill==0.3.8
distro==1.9.0
docker-pycreds==0.4.0
einops==0.8.1
entrypoints==0.4
esdk-obs-python==3.23.12
eval_type_backport==0.2.2
exceptiongroup==1.2.2
executing==1.2.0
fastapi==0.115.11
fastjsonschema==2.21.1
filelock==3.12.2
Flask==3.1.0
fonttools==4.41.1
fqdn==1.5.1
frozenlist==1.5.0
fschat==0.2.36
fsspec==2024.3.1
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.22.0
google-auth-oauthlib==1.0.0
greenlet==3.2.0
grpcio==1.56.2
gunicorn==23.0.0
h11==0.14.0
hjson==3.1.0
httpcore==1.0.7
httpx==0.28.1
huaweicloudsdkcore==3.1.94
huggingface-hub==0.29.2
idna==3.4
ijson==3.3.0
importlib-metadata==6.8.0
ipykernel==6.29.5
ipython==8.14.0
ipython-genutils==0.2.0
ipywidgets==8.1.5
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.18.2
Jinja2==3.1.6
jiter==0.8.2
jmespath==0.10.0
joblib==1.3.1
json5==0.10.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-server==1.24.0
jupyter_client==7.4.9
jupyter_core==5.3.0
jupyter_server_terminals==0.5.3
jupyterlab==3.5.3
jupyterlab-server==2.10.3
jupyterlab_pygments==0.3.0
jupyterlab_widgets==3.0.13
kiwisolver==1.4.4
latex2mathml==3.77.0
lazy-import==0.2.2
lit==16.0.6
lxml==4.9.1
ma-cli==1.2.3
Markdown==3.4.4
markdown-it-py==3.0.0
markdown2==2.5.3
MarkupSafe==2.1.3
matplotlib==3.6.1
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.1.2
mmcv==1.7.1
mock==5.1.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
multiprocess==0.70.16
nbclassic==1.2.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.5.6
networkx==3.1
nh3==0.2.21
ninja==1.11.1.4
notebook==6.5.7
notebook_shim==0.2.4
numpy==1.26.4
nvidia-ml-py==12.570.86
oauthlib==3.2.2
openai==1.65.5
opencv-python==4.7.0.72
overrides==7.7.0
packaging==23.1
pandas==1.5.1
pandocfilters==1.5.1
parso==0.8.3
pathlib2==2.3.7.post1
pexpect==4.8.0
pickleshare==0.7.5
pillow==11.1.0
platformdirs==3.5.1
prettytable==3.7.0
prometheus_client==0.21.1
prompt-toolkit==3.0.38
propcache==0.3.0
protobuf==3.20.1
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyarrow==19.0.1
pyarrow-hotfix==0.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.22
pycryptodome==3.23.0
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.15.1
pyparsing==3.1.0
pytest-runner==5.3.0
python-dateutil==2.8.2
python-json-logger==3.3.0
python-slugify==8.0.4
pytz==2023.3
PyYAML==6.0
pyzmq==25.1.0
referencing==0.36.2
regex==2024.11.6
requests==2.27.1
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rpds-py==0.23.1
rsa==4.9
s3transfer==0.4.2
safetensors==0.5.3
scikit-learn==1.1.3
scipy==1.13.1
semantic-version==2.10.0
Send2Trash==1.8.3
sentencepiece==0.2.0
sentry-sdk==2.22.0
setproctitle==1.3.5
shortuuid==1.0.13
simplejson==3.20.1
six==1.16.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.40
stack-data==0.6.2
starlette==0.46.1
svgwrite==1.4.3
sympy==1.12
tabulate==0.9.0
tenacity==8.2.2
tensorboard==2.12.2
tensorboard-data-server==0.7.1
tensorboard-plugin-wit==1.8.1
terminado==0.18.1
text-unidecode==1.3
threadpoolctl==3.2.0
tiktoken==0.9.0
tinycss2==1.4.0
tokenizers==0.21.0
tomli==2.0.1
torch==2.0.0+cu117
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
transformers==4.49.0
triton==2.0.0
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
tzdata==2025.2
uri-template==1.3.0
urllib3==1.26.16
uvicorn==0.34.0
wandb==0.19.8
wavedrom==2.0.3.post3
wcwidth==0.2.6
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.1.3
widgetsnbextension==4.0.13
XlsxWriter==3.2.3
xxhash==3.5.0
yapf==0.40.1
yarl==1.18.3
zipp==3.16.2
